Modified Login Register

+++ Shared/DTO/Stop/StopDTO.cs
@@ -8,7 +8,7 @@ namespace Shared.DTO
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public decimal Lat { get; set; }
-        public decimal Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
     }
 }
