+++ Shared/DTO/Stop/StopPutDTO.cs
@@ -7,7 +7,7 @@ namespace Shared.DTO
     public class StopPutDTO
     {
         public string? Name { get; set; }
-        public decimal Lat { get; set; }
-        public decimal Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
     }
 }
