+++ WebAPI/Models/Stop.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string? stopName { get; set; }
-        public decimal stopLat {get; set;}
-        public decimal stopLon {get; set;}
+        public double stopLat {get; set;}
+        public double stopLon {get; set;}
 
     }
 }
